fix tag search not working
it seems that if you perform a search, it would try to redirect you to /tag/, but for some reason, even if it's https (ex. https://vx.maid.zone/tag) the server would try to redirect to http (http://vx.maid.zone/tag/), which violates the content-security-policy directive. hopefullythis i